small server refactor
